news_scrapper_api/requirements.txt
# Farmer Advisory System — All Dependencies

feedparser>=6.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
python-multipart>=0.0.9
gtts>=2.4.0
